fix(*): move CredentialsProvider to another folder
