+++ packages/core/src/credentials-provider/CredentialsProvider.ts
