+++ packages/core/src/credentials-provider/DefaultCredentialProvider.ts
