+++ packages/core/src/credentials-provider/index.ts
@@ -1,0 +1,2 @@
+export * from './CredentialsProvider';
+export * from './DefaultCredentialProvider';
